pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "linuxchatbox"
version = "0.2.0"
description = "Native Linux VRChat chatbox for MPRIS media players"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "LinuxChatbox Contributors"}
]
keywords = ["vrchat", "osc", "mpris", "linux", "chatbox", "media"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "PyQt6>=6.0.0",
    "dbus-python>=1.2.0",
    "python-osc>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]

[project.scripts]
linuxchatbox = "linuxchatbox.main:main"

[project.urls]
Homepage = "https://github.com/reskyuu/linuxchatbox"
Repository = "https://github.com/reskyuu/linuxchatbox"
Issues = "https://github.com/reskyuu/linuxchatbox/issues"

[tool.setuptools]
packages = ["linuxchatbox", "linuxchatbox.gui", "linuxchatbox.workers", "linuxchatbox.core"]

[tool.setuptools.package-data]
linuxchatbox = ["resources/*"]
